Make it possible for clients to move around on the map
Currently clients can create a WebSocket connection to the server upon which an entity is created and assigned to them.  Thy can then issue 'commands' that cause the server to update the position of their associated entity.  For example, a client can issue the 'forward' command and then his entity will keep walking in the same direction until either the 'stop_frontal' or the 'backward' command is issued.  At the moment a Map is used to represent all entities in the game and the game state is updated by running a several Systems (as in Entity-Component-System) one after the other.  Right now there are three systems: Networked, IntentToAction and Movement.  The system called 'Networked' only looks at entities having the Controllable component.  Every entity controlled by a client has this component.  This component is meant to represent what the player wants the entity to do without knowing anything about the world or physics.  The second system that runs is called 'IntentToAction' and, as the name is supposed to indicate, its purpose is to to take the player's intents as encoded in Controllable and translate them into an action of their entity.  For example, if the player wants their alter ego to go forward, then IntentToAction will update the entity's Velocity accordingly.  Finally, a system called 'Movement' looks at an entity's velocity and updates their Displacement.

What follows is a list of things that should be done next:

* At the moment entities are not moving around in three dimensions
* the orientation of each entity is fixed (players can't turn around)
* updating an entity's displacement based on how much time elapsed since the last update and the entity's speed